requirements.txt
absl-py==2.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
astunparse==1.6.3
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
fastapi==0.128.0
flatbuffers==25.12.19
gast==0.7.0
google-pasta==0.2.0
grpcio==1.76.0
h11==0.16.0
h5py==3.15.1
idna==3.11
keras==3.13.2
libclang==18.1.1
markdown==3.10.1
markdown-it-py==4.0.0
markupsafe==3.0.3
mdurl==0.1.2
ml-dtypes==0.5.4
namex==0.1.0
numpy==2.4.2
opt-einsum==3.4.0
optree==0.18.0
packaging==26.0
pillow==12.1.0
protobuf==6.33.5
pydantic==2.12.5
pydantic-core==2.41.5
pygments==2.19.2
python-multipart==0.0.22
requests==2.32.5
rich==14.3.2
setuptools==80.10.2
six==1.17.0
starlette==0.50.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
termcolor==3.3.0
typing-extensions==4.15.0
typing-inspection==0.4.2
urllib3==2.6.3
uvicorn==0.40.0
werkzeug==3.1.5
wheel==0.46.3
wrapt==2.1.0
